Use the json module as fallback for simplejson, and
remove the hard dependency on simplejson.

The motivation for this is that simplejson is not required, and
it causes problems with Django 1.5 (see
https://github.com/simplejson/simplejson/issues/37).

For most cases, the speedup of simplejson is not required, and if
the system does not have a C compiler, simplejson is actually slower
than the json module (see https://code.djangoproject.com/ticket/18023).